Add files via upload
This is a basic game that lets you move around a maze using the keyboard. The maze is made of a multidimensional list, which is a list with other lists inside it. Each inner list is a row of the maze, and each item in the inner list is a cell of the maze. The cells can be different things like walls, paths, powerups or exit. The program uses loops and conditions to change and show the maze depending on what you do.